Fixed screen resolution issues.  Tested web build.  Added resource cost and score.

+++ Unity3d/KGS_Defender/Assets/Scripts/Enemy Scripts/EnemyController.cs
@@ -19,7 +19,7 @@ public class EnemyController : MonoBehaviour
     public float attackPower;
     public float attackSpeed;
     public float moveSpeed;
-    public float resourceValue;
+    public int resourceValue;
 
     // Use this for initialization
     void Start ()
